requirements.txt
asgiref==3.8.1
Django==5.1.7
django-countries==7.6.1
django-phonenumber-field==7.2.0
phonenumbers==8.13.25
sqlparse==0.5.3
typing_extensions==4.12.2
